backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15
psycopg2-binary>=2.9
django-filter>=24.2
drf-spectacular>=0.27
drf-spectacular[sidecar]>=0.27
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-dotenv>=1.0
